backend/requirements.txt
fastapi>=0.100.0
sqlmodel>=0.0.14
alembic>=1.13.0
asyncpg>=0.29.0
greenlet>=3.0.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-core>=2.0.0
pydantic-settings>=2.1.0
bcrypt>=4.1.0
python-multipart>=0.0.7
pyjwt[crypto]>=2.8.0
